requirements.txt
requests
python-dotenv
beautifulsoup4
sqlalchemy
mysqlclient
alembic
